refacto(rocksdb): made better use of rocksdb multi get and batch inserts
