+++ Cargo.toml
@@ -10,7 +10,6 @@ members = [
   "crates/pallets/starknet/runtime_api/",
   "crates/primitives/block",
   "crates/primitives/chain-id",
-  "crates/primitives/contract",
   "crates/primitives/convert",
   "crates/primitives/digest-log",
   "crates/primitives/felt",
@@ -218,7 +217,6 @@ pallet-starknet-runtime-api = { path = "crates/pallets/starknet/runtime_api", de
 # Deoxys primtitives
 mp-block = { path = "crates/primitives/block", default-features = false }
 mp-chain-id = { path = "crates/primitives/chain-id", default-features = false }
-mp-contract = { path = "crates/primitives/contract", default-features = false }
 mp-convert = { path = "crates/primitives/convert", default-features = false }
 mp-digest-log = { path = "crates/primitives/digest-log", default-features = false, features = [
   "std",
