+++ crates/client/db/Cargo.toml
@@ -23,15 +23,18 @@ parity-scale-codec = { workspace = true, default-features = true, features = [
 ] }
 sc-client-db = { workspace = true, default-features = true }
 sp-core = { workspace = true }
+cairo-vm = { workspace = true }
+flate2 = { workspace = true }
+indexmap = { workspace = true }
 sp-database = { workspace = true, default-features = true }
 sp-runtime = { workspace = true, default-features = true }
 
 # Deoxys crates
 mp-block = { workspace = true }
-mp-contract = { workspace = true }
 mp-felt = { workspace = true }
 mp-hashers = { workspace = true }
 mp-types = { workspace = true }
+mp-transactions = { workspace = true }
 
 # Starknet crates
 blockifier = { workspace = true, features = ["testing"] }
