+++ crates/client/db/src/storage_handler/contract_data.rs
@@ -1,19 +1,15 @@
-use std::collections::BTreeMap;
-use std::sync::Arc;
-
 use async_trait::async_trait;
 use crossbeam_skiplist::SkipMap;
-use mp_contract::class::StorageContractData;
-use parity_scale_codec::{Decode, Encode};
+use itertools::izip;
 use rocksdb::{IteratorMode, WriteBatchWithTransaction};
-use starknet_api::core::{ClassHash, ContractAddress};
-use tokio::task::JoinSet;
+use starknet_api::core::{ClassHash, ContractAddress, Nonce};
 
+use super::primitives::contract::StorageContractData;
 use super::{DeoxysStorageError, StorageType, StorageView, StorageViewMut, StorageViewRevetible};
 use crate::{Column, DatabaseExt, DeoxysBackend};
 
 #[derive(Default, Debug)]
-pub struct ContractDataViewMut(SkipMap<ContractAddress, StorageContractData>);
+pub struct ContractDataViewMut(SkipMap<ContractAddress, (Option<ClassHash>, Option<Nonce>)>);
 pub struct ContractDataView;
 
 impl StorageView for ContractDataView {
@@ -24,23 +20,23 @@ impl StorageView for ContractDataView {
         let db = DeoxysBackend::expose_db();
         let column = db.get_column(Column::ContractData);
 
-        let tree: BTreeMap<u64, StorageContractData> = match db
-            .get_cf(&column, contract_address.encode())
+        match db
+            .get_cf(&column, bincode::serialize(&contract_address).unwrap())
             .map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData))?
         {
-            Some(bytes) => BTreeMap::decode(&mut &bytes[..])
-                .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?,
-            None => BTreeMap::new(),
-        };
-
-        Ok(tree.last_key_value().map(|(_k, v)| v).cloned())
+            Some(bytes) => Ok(Some(
+                bincode::deserialize::<StorageContractData>(&bytes[..])
+                    .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?,
+            )),
+            None => Ok(None),
+        }
     }
 
     fn contains(&self, contract_address: &Self::KEY) -> Result<bool, DeoxysStorageError> {
         let db = DeoxysBackend::expose_db();
         let column = db.get_column(Column::ContractData);
 
-        match db.key_may_exist_cf(&column, contract_address.encode()) {
+        match db.key_may_exist_cf(&column, bincode::serialize(&contract_address).unwrap()) {
             true => Ok(self.get(contract_address)?.is_some()),
             false => Ok(false),
         }
@@ -48,66 +44,65 @@ impl StorageView for ContractDataView {
 }
 
 impl ContractDataView {
-    pub fn get_at(
+    pub fn get_nonce(self, contract_address: &ContractAddress) -> Result<Option<Nonce>, DeoxysStorageError> {
+        self.get(contract_address)
+            .map(|option| option.map(|contract_data| contract_data.nonce.get().cloned().unwrap_or_default()))
+    }
+
+    pub fn get_class_hash(self, contract_address: &ContractAddress) -> Result<Option<ClassHash>, DeoxysStorageError> {
+        self.get(contract_address).map(|option| {
+            option.and_then(|contract_data| match contract_data.class_hash.get().cloned() {
+                Some(class_hash) => Some(class_hash),
+                None => None,
+            })
+        })
+    }
+
+    pub fn get_nonce_at(
         self,
         contract_address: &ContractAddress,
         block_number: u64,
-    ) -> Result<Option<StorageContractData>, DeoxysStorageError> {
+    ) -> Result<Option<Nonce>, DeoxysStorageError> {
         let db = DeoxysBackend::expose_db();
         let column = db.get_column(Column::ContractData);
 
-        let tree: BTreeMap<u64, StorageContractData> = match db
-            .get_cf(&column, contract_address.encode())
+        let contract_data = match db
+            .get_cf(&column, bincode::serialize(&contract_address).unwrap())
             .map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData))?
         {
-            Some(bytes) => BTreeMap::decode(&mut &bytes[..])
+            Some(bytes) => bincode::deserialize::<StorageContractData>(&bytes[..])
                 .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?,
-            None => BTreeMap::new(),
+            None => StorageContractData::default(),
         };
 
-        Ok(tree.range(..=block_number).next_back().map(|(_, value)| value).cloned())
+        Ok(contract_data.nonce.get_at(block_number).cloned())
     }
-}
 
-impl StorageView for ContractDataViewMut {
-    type KEY = ContractAddress;
-    type VALUE = StorageContractData;
-
-    fn get(&self, contract_address: &Self::KEY) -> Result<Option<Self::VALUE>, DeoxysStorageError> {
-        if let Some(entry) = self.0.get(contract_address) {
-            return Ok(Some(entry.value().clone()));
-        }
-
+    pub fn get_class_hash_at(
+        self,
+        contract_address: &ContractAddress,
+        block_number: u64,
+    ) -> Result<Option<ClassHash>, DeoxysStorageError> {
         let db = DeoxysBackend::expose_db();
         let column = db.get_column(Column::ContractData);
 
-        let tree: BTreeMap<u64, StorageContractData> = match db
-            .get_cf(&column, contract_address.encode())
+        let contract_data = match db
+            .get_cf(&column, bincode::serialize(&contract_address).unwrap())
             .map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData))?
         {
-            Some(bytes) => BTreeMap::decode(&mut &bytes[..])
+            Some(bytes) => bincode::deserialize::<StorageContractData>(&bytes[..])
                 .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?,
-            None => BTreeMap::new(),
+            None => StorageContractData::default(),
         };
 
-        Ok(tree.last_key_value().map(|(_k, v)| v).cloned())
-    }
-
-    fn contains(&self, contract_address: &Self::KEY) -> Result<bool, DeoxysStorageError> {
-        let db = DeoxysBackend::expose_db();
-        let column = db.get_column(Column::ContractData);
-
-        match db.key_may_exist_cf(&column, contract_address.encode()) {
-            true => Ok(self.get(contract_address)?.is_some()),
-            false => Ok(false),
-        }
+        Ok(contract_data.class_hash.get_at(block_number).cloned())
     }
 }
 
 #[async_trait]
 impl StorageViewMut for ContractDataViewMut {
     type KEY = ContractAddress;
-    type VALUE = StorageContractData;
+    type VALUE = (Option<ClassHash>, Option<Nonce>);
 
     fn insert(&self, contract_address: Self::KEY, contract_data: Self::VALUE) -> Result<(), DeoxysStorageError> {
         self.0.insert(contract_address, contract_data);
@@ -116,50 +111,103 @@ impl StorageViewMut for ContractDataViewMut {
 
     async fn commit(self, block_number: u64) -> Result<(), DeoxysStorageError> {
         let db = DeoxysBackend::expose_db();
-        // let column = db.get_column(Column::ContractData);
-        // let (keys, value): (Vec<_>, Vec<_>) = self.0.into_iter().unzip();
-        // let keys_cf = keys.iter().map(|key| (&column, bincode::serialize(key).unwrap()));
-        // let histories = db.multi_get_cf(keys_cf).into_iter().map(|result| {
-        //     result.and_then(|option| match option {
-        //         Some(bytes) => bincode::deserialize(&bytes)
-        //             .map_err(|_|
-        // DeoxysStorageError::StorageDecodeError(StorageType::ContractData)),         None
-        // => todo!(),     })
-        // });
-
-        let mut set = JoinSet::new();
-        for (key, value) in self.0.into_iter() {
-            let db = Arc::clone(db);
-
-            set.spawn(async move {
-                let column = db.get_column(Column::ContractData);
-
-                let Ok(tree) = db.get_cf(&column, key.encode()) else {
-                    return Err(DeoxysStorageError::StorageRetrievalError(StorageType::ContractData));
-                };
-
-                let mut tree: BTreeMap<u64, StorageContractData> = match tree {
-                    Some(bytes) => match BTreeMap::decode(&mut &bytes[..]) {
-                        Ok(tree) => tree,
-                        Err(_) => return Err(DeoxysStorageError::StorageDecodeError(StorageType::ContractData)),
+        let column = db.get_column(Column::ContractData);
+        let (keys, values): (Vec<_>, Vec<_>) = self.0.into_iter().unzip();
+        let keys_cf = keys.iter().map(|key| (&column, bincode::serialize(key).unwrap()));
+        let histories = db
+            .multi_get_cf(keys_cf)
+            .into_iter()
+            .map(|result| {
+                result.map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData)).and_then(
+                    |option| match option {
+                        Some(bytes) => bincode::deserialize::<StorageContractData>(&bytes)
+                            .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData)),
+                        None => Ok(StorageContractData::default()),
                     },
-                    None => BTreeMap::new(),
-                };
+                )
+            })
+            .collect::<Result<Vec<_>, _>>()?;
 
-                tree.insert(block_number, value);
-                if db.put_cf(&column, key.encode(), tree.encode()).is_err() {
-                    return Err(DeoxysStorageError::StorageCommitError(StorageType::ContractData));
-                }
+        let mut batch = WriteBatchWithTransaction::<true>::default();
+        for (key, mut contract_data, (class_hash, nonce)) in izip!(keys, histories, values) {
+            if let Some(class_hash) = class_hash {
+                contract_data.class_hash.push(block_number, class_hash).unwrap();
+            }
 
-                Ok(())
-            });
+            if let Some(nonce) = nonce {
+                contract_data.nonce.push(block_number, nonce).unwrap();
+            }
+
+            batch.put_cf(&column, bincode::serialize(&key).unwrap(), bincode::serialize(&contract_data).unwrap());
         }
+        db.write(batch).map_err(|_| DeoxysStorageError::StorageCommitError(StorageType::ContractData))
 
-        while let Some(res) = set.join_next().await {
-            res.unwrap()?;
+        // let mut set = JoinSet::new();
+        // for (key, value) in self.0.into_iter() {
+        //     let db = Arc::clone(db);
+        //
+        //     set.spawn(async move {
+        //         let column = db.get_column(Column::ContractData);
+        //
+        //         let Ok(tree) = db.get_cf(&column, key.encode()) else {
+        //             return
+        // Err(DeoxysStorageError::StorageRetrievalError(StorageType::ContractData));
+        //         };
+        //
+        //         let mut tree: BTreeMap<u64, StorageContractData> = match tree {
+        //             Some(bytes) => match BTreeMap::decode(&mut &bytes[..]) {
+        //                 Ok(tree) => tree,
+        //                 Err(_) => return
+        // Err(DeoxysStorageError::StorageDecodeError(StorageType::ContractData)),
+        //             },
+        //             None => BTreeMap::new(),
+        //         };
+        //
+        //         tree.insert(block_number, value);
+        //         if db.put_cf(&column, key.encode(), tree.encode()).is_err() {
+        //             return
+        // Err(DeoxysStorageError::StorageCommitError(StorageType::ContractData));         }
+        //
+        //         Ok(())
+        //     });
+        // }
+        //
+        // while let Some(res) = set.join_next().await {
+        //     res.unwrap()?;
+        // }
+        //
+        // Ok(())
+    }
+}
+
+impl StorageView for ContractDataViewMut {
+    type KEY = ContractAddress;
+    type VALUE = StorageContractData;
+
+    fn get(&self, contract_address: &Self::KEY) -> Result<Option<Self::VALUE>, DeoxysStorageError> {
+        let db = DeoxysBackend::expose_db();
+        let column = db.get_column(Column::ContractData);
+
+        match db
+            .get_cf(&column, bincode::serialize(&contract_address).unwrap())
+            .map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData))?
+        {
+            Some(bytes) => Ok(Some(
+                bincode::deserialize::<StorageContractData>(&bytes[..])
+                    .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?,
+            )),
+            None => Ok(None),
         }
+    }
 
-        Ok(())
+    fn contains(&self, contract_address: &Self::KEY) -> Result<bool, DeoxysStorageError> {
+        let db = DeoxysBackend::expose_db();
+        let column = db.get_column(Column::ContractData);
+
+        match db.key_may_exist_cf(&column, bincode::serialize(&contract_address).unwrap()) {
+            true => Ok(self.get(contract_address)?.is_some()),
+            false => Ok(false),
+        }
     }
 }
 
@@ -174,13 +222,20 @@ impl StorageViewRevetible for ContractDataViewMut {
         for data in iterator {
             let (key, value) =
                 data.map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData))?;
-            let mut tree: BTreeMap<u64, ClassHash> = BTreeMap::decode(&mut &value[..])
+            let mut contract_data = bincode::deserialize::<StorageContractData>(&value[..])
                 .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?;
 
-            tree.retain(|&k, _| k <= block_number);
+            contract_data.class_hash.revert_to(block_number);
+            contract_data.nonce.revert_to(block_number);
 
-            db.put_cf(&column, key, tree.encode())
-                .map_err(|_| DeoxysStorageError::StorageRevertError(StorageType::ContractData, block_number))?;
+            match (contract_data.class_hash.is_empty(), contract_data.nonce.is_empty()) {
+                (true, true) => db
+                    .delete(key)
+                    .map_err(|_| DeoxysStorageError::StorageRevertError(StorageType::ContractData, block_number))?,
+                _ => db
+                    .put_cf(&column, key, bincode::serialize(&contract_data).unwrap())
+                    .map_err(|_| DeoxysStorageError::StorageRevertError(StorageType::ContractData, block_number))?,
+            }
         }
 
         Ok(())
@@ -188,25 +243,49 @@ impl StorageViewRevetible for ContractDataViewMut {
 }
 
 impl ContractDataViewMut {
+    pub fn insert_nonce(&self, contract_address: ContractAddress, nonce: Nonce) -> Result<(), DeoxysStorageError> {
+        self.insert(contract_address, (None, Some(nonce)))
+    }
+
+    pub fn insert_class_hash(
+        &self,
+        contract_address: ContractAddress,
+        class_hash: ClassHash,
+    ) -> Result<(), DeoxysStorageError> {
+        self.insert(contract_address, (Some(class_hash), None))
+    }
+
     pub fn commit_sync(self, block_number: u64) -> Result<(), DeoxysStorageError> {
         let db = DeoxysBackend::expose_db();
         let column = db.get_column(Column::ContractData);
+        let (keys, values): (Vec<_>, Vec<_>) = self.0.into_iter().unzip();
+        let keys_cf = keys.iter().map(|key| (&column, bincode::serialize(key).unwrap()));
+        let histories = db
+            .multi_get_cf(keys_cf)
+            .into_iter()
+            .map(|result| {
+                result.map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData)).and_then(
+                    |option| match option {
+                        Some(bytes) => bincode::deserialize::<StorageContractData>(&bytes)
+                            .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData)),
+                        None => Ok(StorageContractData::default()),
+                    },
+                )
+            })
+            .collect::<Result<Vec<_>, _>>()?;
 
-        for (key, value) in self.0.into_iter() {
-            let mut tree: BTreeMap<u64, StorageContractData> = match db
-                .get_cf(&column, key.encode())
-                .map_err(|_| DeoxysStorageError::StorageRetrievalError(StorageType::ContractData))?
-            {
-                Some(bytes) => BTreeMap::decode(&mut &bytes[..])
-                    .map_err(|_| DeoxysStorageError::StorageDecodeError(StorageType::ContractData))?,
-                None => BTreeMap::new(),
-            };
+        let mut batch = WriteBatchWithTransaction::<true>::default();
+        for (key, mut contract_data, (class_hash, nonce)) in izip!(keys, histories, values) {
+            if let Some(class_hash) = class_hash {
+                contract_data.class_hash.push(block_number, class_hash).unwrap();
+            }
 
-            tree.insert(block_number, value);
-            db.put_cf(&column, key.encode(), tree.encode())
-                .map_err(|_| DeoxysStorageError::StorageCommitError(StorageType::ContractData))?;
+            if let Some(nonce) = nonce {
+                contract_data.nonce.push(block_number, nonce).unwrap();
+            }
+
+            batch.put_cf(&column, bincode::serialize(&key).unwrap(), bincode::serialize(&contract_data).unwrap());
         }
-
-        Ok(())
+        db.write(batch).map_err(|_| DeoxysStorageError::StorageCommitError(StorageType::ContractData))
     }
 }
