+++ crates/client/db/src/storage_handler/primitives/mod.rs
@@ -1,0 +1,2 @@
+pub mod contract;
+pub mod contract_class;
