+++ crates/client/rpc/src/utils/blockifier_state_adapter.rs
@@ -53,8 +53,8 @@ impl StateReader for BlockifierStateAdapter {
     fn get_nonce_at(&self, contract_address: ContractAddress) -> StateResult<Nonce> {
         match self.nonce_update.get(&contract_address) {
             Some(nonce) => Ok(*nonce),
-            None => match storage_handler::contract_data().get_at(&contract_address, self.block_number) {
-                Ok(Some(contract_data)) => Ok(contract_data.nonce),
+            None => match storage_handler::contract_data().get_nonce_at(&contract_address, self.block_number) {
+                Ok(Some(nonce)) => Ok(nonce),
                 Ok(None) => Ok(Nonce::default()),
                 Err(_) => Err(StateError::StateReadError(format!(
                     "Failed to retrieve nonce for contract {}",
@@ -67,8 +67,8 @@ impl StateReader for BlockifierStateAdapter {
     fn get_class_hash_at(&self, contract_address: ContractAddress) -> StateResult<ClassHash> {
         match self.class_hash_update.get(&contract_address).cloned() {
             Some(class_hash) => Ok(class_hash),
-            None => match storage_handler::contract_data().get_at(&contract_address, self.block_number) {
-                Ok(Some(contract_data)) => Ok(contract_data.class_hash),
+            None => match storage_handler::contract_data().get_class_hash_at(&contract_address, self.block_number) {
+                Ok(Some(class_hash)) => Ok(class_hash),
                 _ => Err(StateError::StateReadError(format!(
                     "failed to retrive class hash for contract address {}",
                     contract_address.0.0
