+++ crates/client/rpc/src/utils/execution.rs
@@ -69,8 +69,7 @@ pub fn call_contract(
     block_context: &BlockContext,
 ) -> Result<Vec<Felt252Wrapper>, ()> {
     // Get class hash
-    let class_hash =
-        storage_handler::contract_data().get(&address).map_err(|_| ())?.map(|contract_data| contract_data.class_hash);
+    let class_hash = storage_handler::contract_data().get_class_hash(&address).map_err(|_| ())?;
 
     let entrypoint = CallEntryPoint {
         class_hash,
