+++ crates/client/rpc/src/utils/transaction.rs
@@ -2,8 +2,8 @@ use blockifier::execution::contract_class::ClassInfo;
 use blockifier::transaction::transaction_execution as btx;
 use jsonrpsee::core::RpcResult;
 use mc_db::storage_handler;
+use mc_db::storage_handler::primitives::contract_class::StorageContractClassData;
 use mc_db::storage_handler::StorageView;
-use mp_contract::class::StorageContractClassData;
 use starknet_api::hash::StarkFelt;
 use starknet_api::transaction::{Transaction, TransactionHash};
 use starknet_ff::FieldElement;
