+++ crates/client/sync/Cargo.toml
@@ -56,7 +56,6 @@ bitvec = { workspace = true }
 bonsai-trie = { workspace = true }
 mc-db = { workspace = true }
 mp-block = { workspace = true }
-mp-contract = { workspace = true }
 mp-convert = { workspace = true }
 mp-felt = { workspace = true }
 mp-hashers = { workspace = true }
