+++ crates/client/sync/src/fetch/fetchers.rs
@@ -4,9 +4,9 @@ use std::sync::Arc;
 
 use itertools::Itertools;
 use mc_db::storage_handler;
+use mc_db::storage_handler::primitives::contract_class::{ContractClassData, ContractClassWrapper};
 use mc_db::storage_handler::StorageView;
 use mp_block::DeoxysBlock;
-use mp_contract::class::{ContractClassData, ContractClassWrapper};
 use mp_convert::state_update::ToStateUpdateCore;
 use sp_core::H160;
 use starknet_api::core::ClassHash;
