+++ crates/pallets/starknet/Cargo.toml
@@ -20,7 +20,6 @@ targets = ["x86_64-unknown-linux-gnu"]
 # Deoxys primitives
 mc-db = { workspace = true }
 mp-block = { workspace = true }
-mp-contract = { workspace = true }
 mp-digest-log = { workspace = true }
 mp-felt = { workspace = true, features = ["parity-scale-codec", "serde"] }
 mp-genesis-config = { workspace = true, default-features = false }
@@ -96,7 +95,6 @@ std = [
   # Other third party dependencies
   "dep:reqwest",
   "mp-block/std",
-  "mp-contract/std",
   "mp-simulations/std",
   "parity-scale-codec/std",
   "starknet-core/std",
