--- crates/primitives/contract/Cargo.toml
@@ -1,46 +1,0 @@
-[package]
-authors.workspace = true
-description = "The starknet contract"
-edition.workspace = true
-license = "MIT"
-name = "mp-contract"
-repository.workspace = true
-version.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-blockifier = { workspace = true }
-log = { workspace = true, default-features = true }
-starknet-core = { workspace = true }
-starknet_api = { workspace = true }
-
-# Optionals
-anyhow = { workspace = true, optional = true }
-mp-convert = { workspace = true, optional = true }
-mp-felt = { workspace = true }
-parity-scale-codec = { workspace = true, features = [
-  "derive",
-], optional = true }
-scale-info = { workspace = true, features = ["derive"], optional = true }
-serde = { workspace = true }
-
-[features]
-default = ["std"]
-parity-scale-codec = [
-  "dep:parity-scale-codec",
-  "starknet_api/parity-scale-codec",
-]
-scale-info = [
-  "blockifier/scale-info",
-  "dep:scale-info",
-  "starknet_api/scale-info",
-]
-std = [
-  # Optionals
-  "anyhow/std",
-  "mp-convert/std",
-  "parity-scale-codec?/std",
-  "scale-info?/std",
-]
