+++ crates/primitives/convert/src/lib.rs
@@ -1,4 +1,3 @@
 #[cfg(feature = "std")]
-pub mod contract;
 pub mod field_element;
 pub mod state_update;
